requirements.txt
httpx
asyncio
authlib
rich
pyyaml
